First draft of utility to extract opcodes from the Intel Instruction Set Reference Volume 2B, Appendix B.
I cannot guarantee the accuracy of the output due to the fact that the
original document was not formatted for automated parsing, hence the
extraction process is clumsy and error-prone.  In addition, the original
document has various typos; while I added logic to workaround all the
errors I am aware of, it is possible others exist.

Approximate time spent: 25 hours.

Still to-do:
  * Extract floating point instructions.
  * Output to XML.