Interface for interacting with the ptrace(2) process tracing facility.
Encapsulated all of the state necessary to transparently trace another
process.  That is, intercepts signals for the traced process, filters the
SIGTRAP signals generated by the act of tracing, and forwards the
remainder to the traced process as usual.

Approximate time spent: 10 hours

TODO: Investigate why grand-children die with SIGTRAP.  I suspect that
      tracing is being inherited hence when the grandchild traps its
      parent is seeing the SIGTRAP in the WSTOPSIG(status) from wait(2).
      This is reproducable by running dynprof on /bin/sh and trying to
      run some non-builtin command like ls.