Import opcode list automatically extracted from Intel's IA-32 Software Developer's Manual, publication #25366714, appendix 3 (Instruction Formats and Encodings).
Imported as a vendor import even though it was generated using our own
extract.pl script so that manual edits can be easily merged after future
imports.  That is to say, the automatically-generated opcode list should
always be imported on this vendor branch.  After each import, hand edits
made to the list can the be merged using 'cvs checkout -j'.