Work-in-progress to support differentiating memory regions within the profiled process as well a tracing multiple processes.  The former will be necessary for profiling runtime-generated code (e.g. from Java JIT compilers).  The latter will be necessary for profiling daemons or other processes which fork children to perform work (e.g. apache).
In addition, if we know a memory region is read-only, then we can cache
the contents of that region rather than having to request it from the
kernel after every instruction.  This eliminated context switch will be
replaced with the kernel call to fetch the profiled process' memory map
so there will be no change in profiling throughput.

Approximate time spent:	4 hours