Add an extra layer of indirection between the profiler and the process control implementation.  This is primarilly for the sake of portability. The intent is to allow different platforms to use whatever facilities they have to implement the process control functionality required by the profiler.
The organization is now that state of the traced target is described by a
target_t object.  This target may represent an entire process or a single
thread in the process (depending on the platform's capabilities).  A
region_list_t object now describes all of the regions in the process
VM space of a target.  One or more targets may all share a single
region_list_t, for example, if each target is a thread, then multiple
threads have the same VM space so they would all share the same
region_list_t.

The point is that the platform determines the representation of what a
traceable target is.  The profiler only cares that the platform presents
the API it needs to collect the dynamic trace of the target's (or group of
targets') execution.

The goal is to be able to trace all of the threads in a process and
possibly to include any children of the process in the trace.  Currently,
the profiler cannot do that, but that is the direction I'm aiming in.

While I was at it, I removed the (failed) attempt to mmap(2) /proc/XXX/mem
in order to read the profiled's process's VM space without as many context
switches.

Approximate time spent: 14 hours